Dent0 - Dental clinic Management System made easy
Observing some of the basic public and private services our country has to offer we have come up with a management system for Dental clinics. Although unfortunately a majority of people do not pay attention to their health until it becomes too late, we have tried to make the process of finding a dentist and making an appointment with them much more interacting and doable with a few clicks of a button.  :D

The goal of this project is to develop a web-based dental management system that makes it easy for patients to find new dentists and schedule appointments with them. The system will allow patients to search for dentists based on location, availability, and other criteria, and will provide detailed information about each dentist, including their qualifications, ratings, and reviews.

For Dentists, the system will provide an easy way to manage their appointments, patients information and also can be used to manage the billing process. Dentists will also be able to update their availability and view their schedule through the system which would be a huge help for them, especially when it comes to private clinics where things are handled using the “old ways”.
The system will also be able to handle multiple practices and dental clinics, so dental practices and clinics will be able to use it to manage multiple locations and providers.